Move object_type_identifier() & object_type_name() to utilities.object_types
