+++ netbox/netbox/tables/columns.py
@@ -18,9 +18,9 @@ from django_tables2.columns import library
 from django_tables2.utils import Accessor
 
 from extras.choices import CustomFieldTypeChoices
+from utilities.object_types import object_type_identifier, object_type_name
 from utilities.permissions import get_permission_for_model
 from utilities.templatetags.builtins.filters import render_markdown
-from utilities.utils import object_type_identifier, object_type_name
 from utilities.views import get_viewname
 
 __all__ = (
